Establish a TCP connection on each collect call
This allows the exporter to be resilient to mcrouter restarts,
and avoid segfaulting if mcrouter is not launched before it.

The downside is not re-using any existing TCP connection, leading
to a socket in TIME_WAIT after each collect call. This is not a big
issue on OS in which these sockets can be reused quickly after
a few seconds in specific cases like this one. More info for Linux:

https://vincent.bernat.ch/en/blog/2014-tcp-time-wait-state-linux

The final code looks cleaner and less error prone, so connections
re-use may be enabled/added only if really needed in the future.

PR: #4