Fix per-server tko metric (#13)
In our deployment of the exporter the per-server tko metrics are
always zero. I digged a bit in the issue and I found that "stats servers"
adds two flags, 'soft_tko' and 'hard_tko', but it doesn't seem to have
any "tko" reply per-server metric. When I wrote the first patch I was
probably confused by the docs that talk about a "tko" reply, that is
not reported by mcrouter as far as I can see.

The old per-server tko metric is replaced by two Gauge metrics, one
representing the flag soft_tko and the other one the hard_tko.
I used the values 0/1 to indicate if a server is marked as soft/hard_tko,
it should be understandable but I am open to suggestions.